Add Microserver
Microserver is a plugin system for building Microservices on top of plain old Spring or Spring Boot. It helps solve the problem of abstracting and extracting infrastructure code from common libraries, by providing a controlled & extensible plugin mechanism. Creating a plugin is as simple as creating a Service, and plugins can be used simply by adding them to the classpath.