Removed Faux Pas from Functional Programming section
This library is not related to functional programming. Having it in this section might lead learners of functional programming to think that error propagation/handling via runtime exceptions is the way to go. It is not. Even worse, sneaky throw is a hack/corruption of the type system.
A key goal of functional programming is maximize safety, this library has the opposite goal by encouraging unsafe functions.

In functional programming one would rather use sum types (or at least checked exceptions) for error handling.