Added DatumBox Machine Learning framework
The main focus of the framework is to include a large number of machine learning algorithms & statistical methods and to be able to handle large sized datasets. It uniquely containes various pre-trained models, making it easy to include and use in existing projects. Like detections of topics, spam, adult content, language, gender, sentiment, cemmerials, etc.