Added NoException and removed Faux Pas
This change replaces Faux Pas, which is already listed, with NoException. NoException is a superset of Faux Pas functionality with support for all functional interfaces and with complete documentation. NoException is also a superset of functionality found in other similar libraries (listed on NoException site). Arguably, NoException also has a cleaner API.

Category was changed from Development to the more appropriate Functional Programming.

Disclaimer: I am the author of NoException.