Add DJL
DJL (Deep Java Library) is an open-source library for Deep Learning. It
is engine agnostic so it can use the same native libraries that power
the most popular python deep learning frameworks while allowing you to switch
between them without code changes. It also focuses on ease of use to
make deep learning as easy as possible for Java developers.