Added the DSAA book to "Influential Books" section
The official homepage for the paper (and PDF) version of the book Data Structures & Algorithm Analysis by Clifford A. Shaffer.
The paper/PDF version is no longer actively maintained or updated.The most recent version is Edition 3.2.0.10, dated March 28, 2013. But it is still a great resource, specially for educational use.