Added "DevProjects" under "Miscellaneous"
I added DevProjects to miscellaneous because I've been using the website's projects to practice Java. I'm quite new to Java, so I focus on the easier ones. I personally find the solutions and discussions helpful and think building Java projects is a fun way to learn Java.

If you think it might be a good resource for the community, I think it'd be a great addition!