Add Micrometer and Micrometer Tracing
Widely recommended, regardless of personal opinion:
Micrometer is the de-facto set of metrics/observability libraries
on the JVM. The biggest frameworks/libraries are
instrumenting themselves using Micrometer and Micrometer Tracing,
a few of them: Spring, Quarkus, Micronaut, Hibernate, OpenFeign,
HikariCP, etc. (this list is very long).

Highly discussed in the community due to its innovative nature:
Please check the `micrometer` tag on StackOverflow, Micrometer Slack,
content of Java-related conferences, the documentation of
the Observation API and look around tech blogs.
Micrometer and Micrometer Tracing has been very popular and
highly discussed.

Absolutely unique in its approach and function:
Micrometer was the first library that supported dimensional metrics on
the JVM and as far as I know there are no other libraries that have
a general Observability API as of today. Also, as far as I know,
there are no tracing facades other than Micrometer Tracing on the JVM
(except Spring Cloud Sleuth which is its predecessor).

Niche product that fills a gap: see the previous point.

They are open source, licensed under Apache 2.0, documented in English.

Self-promotion is frowned upon and viewed critically,
but your suggestion will of course be approved if the criteria match:
I'm one of the maintainers of both projects,
so I guess this is self promotion?